+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - EntityTablePanel.SHOW_REFRESH_PROGRESS_BAR renamed REFRESH_PROGRESS_BAR, now true by default.
 - EntitySearchField bug fixed, no longer updates the background color when disabled.
 - EntitySearchField.searchOnFocusLost now final, searchOnFocusLost() removed.
+- EntityEditComponentPanel.InputFocus.transferOnEnter() removed.
 
 ## 0.18.25
 ### is.codion.common.db